Upsert action: utilization of input binary attachment.
