+++ lib/actions/createObject.js
@@ -1,13 +1,13 @@
 'use strict';
 
 const _ = require('lodash');
-const requestPromise = require('request-promise');
 const jsforce = require('jsforce');
 const messages = require('elasticio-node').messages;
 const debug = require('debug')('salesforce:createObject');
 const SalesforceEntity = require('../entry.js').SalesforceEntity;
 const MetaLoader = require('../helpers/metaLoader');
 const common = require('../common.js');
+const attachment = require('../helpers/attachment.js');
 
 exports.objectTypes = async function objectTypes(configuration) {
   const metaLoader = new MetaLoader(configuration, this);
@@ -35,44 +35,7 @@ exports.process = async function createObject(msg, configuration) {
 
   debug("Creating message body: ", msg.body);
 
-  let binField;
-  let attachment;
-  for (attachment in msg.attachments) break;
-
-  // If there is an attachment then the binary data field is overwritten.
-  if (attachment) {
-    attachment = msg.attachments[attachment];
-
-    console.log("Found an attachment from the previous component.");
-
-    if (configuration.utilizeAttachment) {
-      const metaLoader = new MetaLoader(configuration, this, sfConn);
-      const objectFields = await metaLoader.getObjectFieldsMetaData();
-
-      binField = objectFields.find(field => field.type === "base64");
-
-      if (binField) {
-        console.log("Preparing the attachment...");
-
-        const optsDownload = {
-          uri: attachment['url'],
-          method: 'GET',
-          resolveWithFullResponse: true,
-          encoding: null
-        }
-        const response = await requestPromise(optsDownload);
-        msg.body[binField.name] = Buffer.from(response.body).toString('base64');
-
-        if (attachment["content-type"] && objectFields.find(field => field.name === "ContentType"))
-          msg.body.ContentType = attachment["content-type"];
-
-      } else {
-        console.log(`${configuration.sobject} object does not have a field for binary data. The attachment is discarded.`);
-      }
-    } else {
-      console.log("The attachment will be discarded.");
-    }
-  }
+  const binaryField = await attachment.prepareBinaryData(msg, configuration, sfConn);
 
   console.log("Sending request to SalesForce...");
 
@@ -83,8 +46,8 @@ exports.process = async function createObject(msg, configuration) {
     console.log(`${configuration.sobject} has been successfully created (ID = ${response.id}).`);
     msg.body.id = response.id;
 
-    if (binField)
-      delete msg.body[binField.name];
+    if (binaryField)
+      delete msg.body[binaryField.name];
 
     return messages.newMessageWithBody(msg.body);
   } catch (err) {
