+++ lib/actions/upsert.js
@@ -4,6 +4,7 @@ const lookup = require('./lookup');
 const debug = require('debug')('salesforce:upsert');
 const common = require('../common.js');
 const MetaLoader = require('../helpers/metaLoader');
+const attachment = require('../helpers/attachment.js');
 
 /**
  * This function will return a metamodel description for a particular object
@@ -34,8 +35,7 @@ module.exports.objectTypes = async function getObjectTypes(configuration) {
  * @param message
  * @param configuration
  */
-module.exports.process = function upsertObject(message, configuration) {
-  const self = this;
+module.exports.process = async function upsertObject(message, configuration) {
   configuration.object = configuration.sobject;
   console.log('Starting upsertObject');
   const conn = new jsforce.Connection({
@@ -54,13 +54,15 @@ module.exports.process = function upsertObject(message, configuration) {
     this.emit('updateKeys', {oauth: res});
   });
 
+  await attachment.prepareBinaryData(message, configuration, conn);
+
   if (message.body.Id) {
     console.log('Upserting sobject=%s by internalId', configuration.sobject, message.body.Id);
     debug('Upserting %s by internalId data: %j', configuration.sobject, message.body.Id, message);
     return conn.sobject(configuration.sobject).update(message.body)
         .then(() => {
           configuration.lookupField = 'Id';
-          return lookup.process.call(self, {body:{Id: message.body.Id}},
+          return lookup.process.call(this, {body:{Id: message.body.Id}},
               configuration);
         });
   } else {
